refactor: remove filter subcomponents
they don't take up many lines of code so i'll just put them all in the same space
