--- src/components/Disclaimer.js
@@ -1,17 +1,0 @@
-// import './Disclaimer.css';
-import React from "react";
-
-function Disclaimer() {
-  return (
-    <div id="disclaimer">
-      <p>
-        <i>
-          Disclaimer: This website is not an official UC Merced product. Always
-          consult your advisor before making course selection-related decisions.
-        </i>
-      </p>
-    </div>
-  );
-}
-
-export default Disclaimer;
