+++ src/components/Filters.js
@@ -1,7 +1,5 @@
 import React from "react";
-import ChecklistFilter from "./ChecklistFilter.js";
-import UpdateDate from "./UpdateDate.js";
-import Disclaimer from "./Disclaimer.js";
+import { useState, useEffect } from "react";
 import "./Filters.css";
 
 /**
@@ -25,17 +23,71 @@ const badgeList = [
 ];
 
 function Filters(props) {
+  const [date, setDate] = useState("");
+
+  useEffect(() => {
+    async function fetchDate() {
+      try {
+        const response = await fetch(
+          "https://thevanadium.pythonanywhere.com/update-date"
+        );
+        const data = await response.text();
+        console.log(data);
+        setDate(
+          new Date(data).toLocaleDateString("en-US", {
+            year: "numeric",
+            month: "long",
+            day: "numeric",
+          })
+        );
+      } catch (error) {
+        console.error("Error:", error);
+      }
+    }
+    fetchDate();
+  }, []);
+
   return (
     <div id="filters">
       <h2>Filters</h2>
-      <ChecklistFilter
-        checklistItems={badgeList}
-        update={(value) => {
-          props.updateFilters(value);
-        }}
-      ></ChecklistFilter>
-      <UpdateDate></UpdateDate>
-      <Disclaimer></Disclaimer>
+      <div id="badgeList">
+        <p>Badges</p>
+        {badgeList.map((item) => {
+          let urlItem = item
+            .toLowerCase()
+            .replace(/ /g, "-")
+            .replace(/the-|of-/g, "");
+          return (
+            <div key={urlItem} id={item}>
+              <input
+                key={urlItem + "input"}
+                type="checkbox"
+                id={item}
+                name={item}
+                value={urlItem}
+                onChange={(event) => {
+                  props.updateFilters(event.target.value);
+                }}
+              />
+              <label key={urlItem + "label"} htmlFor={item}>
+                {item}
+              </label>
+            </div>
+          );
+        })}
+      </div>
+      <div id="update-date">
+        <p>Last Updated: {date ? date : <i>Loading...</i>}</p>
+      </div>
+      <div id="disclaimer">
+        <p>
+          <i>
+            Disclaimer: This website is not an official UC Merced product.
+            Always consult your advisor before making course selection-related
+            decisions.
+          </i>
+        </p>
+      </div>
     </div>
   );
 }
